+++ evaluator/TA/volatility_evaluator.py
@@ -5,7 +5,6 @@ from evaluator.TA.TA_evaluator import VolatilityEvaluator
 class ATRVolatilityEvaluator(VolatilityEvaluator):
     def __init__(self):
         super().__init__()
-        self.enabled = False
 
     def eval_impl(self):
         pass
@@ -15,7 +14,6 @@ class ATRVolatilityEvaluator(VolatilityEvaluator):
 class MassIndexVolatilityEvaluator(VolatilityEvaluator):
     def __init__(self):
         super().__init__()
-        self.enabled = False
 
     def eval_impl(self):
         pass
@@ -24,7 +22,6 @@ class MassIndexVolatilityEvaluator(VolatilityEvaluator):
 class ChaikinVolatilityEvaluator(VolatilityEvaluator):
     def __init__(self):
         super().__init__()
-        self.enabled = False
 
     def eval_impl(self):
         pass