+++ evaluator/abstract_evaluator.py
@@ -28,6 +28,7 @@ class AbstractEvaluator:
     # Used to provide the global config
     def set_config(self, config):
         self.config = config
+        self.enabled = self.is_enabled(False)
 
     # Symbol is the cryptocurrency symbol
     def set_symbol(self, symbol):
@@ -83,9 +84,25 @@ class AbstractEvaluator:
     def eval_impl(self) -> None:
         raise NotImplementedError("Eval_impl not implemented")
 
+    # explore up to the 1st parent
     @classmethod
     def get_is_dispatcher_client(cls):
-        return DispatcherAbstractClient in cls.__bases__
+        if DispatcherAbstractClient in cls.__bases__:
+            return True
+        else:
+            for base in cls.__bases__:
+                if DispatcherAbstractClient in base.__bases__:
+                    return True
+        return False
+    
+    @classmethod
+    def get_parent_evaluator_classes(cls, higher_parent_class_limit=None):
+        classes = []
+        limit_class = higher_parent_class_limit if higher_parent_class_limit else AbstractEvaluator
+        for class_type in cls.mro():
+            if limit_class in class_type.mro():
+                classes.append(class_type)
+        return classes
 
     def set_eval_note(self, new_eval_note):
         if self.eval_note == START_PENDING_EVAL_NOTE:
@@ -97,3 +114,13 @@ class AbstractEvaluator:
             self.eval_note = -1
         else:
             self.eval_note += new_eval_note
+
+    def is_enabled(self, default):
+        if self.config[CONFIG_EVALUATOR] is not None:
+            if self.get_name() in self.config[CONFIG_EVALUATOR]:
+                return self.config[CONFIG_EVALUATOR][self.get_name()]
+            else:
+                for parent in self.__class__.mro():
+                    if parent.__name__ in self.config[CONFIG_EVALUATOR]:
+                        return self.config[CONFIG_EVALUATOR][parent.__name__]
+                return default