+++ tests/test_start_bot.py
@@ -1,19 +1,21 @@
 import time
 
 from config.cst import *
-from bot import Crypto_Bot
+from cryptobot import CryptoBot
+from tests.test_utils.config import load_test_config
 
 
 def test_create_bot():
     # launch a bot
-    bot = Crypto_Bot()
-    bot.create_exchange_traders()
+    config = load_test_config()
+    bot = CryptoBot(config)
     bot.stop_threads()
 
 
 def test_run_bot():
     # launch a bot
-    bot = Crypto_Bot()
+    config = load_test_config()
+    bot = CryptoBot(config)
     bot.time_frames = [TimeFrames.ONE_MINUTE]
     bot.create_exchange_traders()
     bot.create_evaluation_threads()