+++ trading/trader/trader_simulator.py
@@ -21,8 +21,11 @@ class TraderSimulator(Trader):
         else:
             return False
 
-    def create_order(self, order_type, symbol, quantity, price=None, stop_price=None, linked_to=None):
-        self.logger.info("Order creation : {0} | {1} | Price : {2}".format(symbol, order_type, price))
+    def create_order(self, order_type, symbol, current_price, quantity, price=None, stop_price=None, linked_to=None):
+        self.logger.info("Order creation : {0} | {1} | Price : {2} | Quantity : {3}".format(symbol,
+                                                                                            order_type,
+                                                                                            price,
+                                                                                            quantity))
 
         # create new order instance
         order_class = OrderConstants.TraderOrderTypeClasses[order_type]
@@ -34,13 +37,14 @@ class TraderSimulator(Trader):
         else:
             order_notifier = linked_to.get_order_notifier()
 
-        order.new(order_type, symbol, quantity, price, stop_price, order_notifier)
+        order.new(order_type, symbol, current_price, quantity, price, stop_price, order_notifier)
+
+        # update the availability of the currency in the portfolio
+        with self.portfolio as pf:
+            pf.update_portfolio_available(order, is_new_order=True)
 
         # notify order manager of a new open order
         self.order_manager.add_order_to_list(order)
-
-        # update the availability of the currency in the portfolio
-        self.portfolio.update_portfolio_available(order, True)
 
         # if this order is linked to another (ex : a sell limit order with a stop loss order)
         if linked_to is not None: